requirements.txt
# Web Framework
streamlit>=1.10.0

# HTTP requests
requests>=2.28.0

# MongoDB driver
pymongo>=4.0.0

# FAISS (CPU version)
faiss-cpu>=1.7.2

# Embeddings & LLMs
sentence-transformers>=2.2.2
transformers>=4.41.0
torch>=1.11.0

# Utilities
numpy>=1.23.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0